// app-nievabus/interfaces/viajes/IViajes.ts
// Generated by https://quicktype.io

export interface IViajes {
    referencia:  string;
    descripcion: string;
    origen:      string;
    destino:     string;
    hora_inicio: string;
    ida_vuelta:  string;
    dniCliente:  string;
    dniEmpleado: string;
    imagen:      string;
    precio:      string;
}